% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\docType{data}
\name{egoitz_colors}
\alias{egoitz_colors}
\title{color_palette}
\format{
An object of class \code{character} of length 31.
}
\usage{
egoitz_colors
}
\description{
Some preconfigured color palettes, useful for plots
}
\author{
Egoitz Carral
}
\keyword{datasets}
