% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_variable_analysis.R
\name{continuous_variable_analysis}
\alias{continuous_variable_analysis}
\title{Continuous Variable Analysis}
\usage{
continuous_variable_analysis(
  var1,
  media = TRUE,
  mediana = FALSE,
  moda = FALSE,
  dt = TRUE,
  varianza = TRUE,
  rango = FALSE,
  RIC = TRUE,
  grafico = TRUE,
  simetria = TRUE,
  curtosis = TRUE,
  ...
)
}
\arguments{
\item{var1}{La variable a analizar (numérica y continua)}

\item{media}{Valor booleano, calcular o no la media}

\item{mediana}{Valor booleano, calcular o no la mediana}

\item{moda}{Valor booleano, calcular o no la moda}

\item{dt}{Valor booleano, calcular o no la desviación típica}

\item{varianza}{Valor booleano, calcular o no la varianza}

\item{rango}{Valor booleano, calcular o no el rango}

\item{RIC}{Valor booleano, calcular o no el rangointercuartil}

\item{grafico}{Valor booleano, dibujar o no un histograma}

\item{simetria}{Valor booleano, calcular o no la simetria}

\item{curtosis}{Valor booleano, calcular o no la curtosis}
}
\description{
Continuous Variable Analysis
}
\author{
Egoitz Carral
}
