% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{egoitz_color_pal}
\alias{egoitz_color_pal}
\title{Return function to interpolate a palette}
\usage{
egoitz_color_pal(palette = "blue_green_brown", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in egoitz_color_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a palette
}
