% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{egoitz_cols}
\alias{egoitz_cols}
\title{Function to extract egoitz_colors colors as hexadecimal codes}
\usage{
egoitz_cols(...)
}
\arguments{
\item{...}{Character names of egoitz_colors}
}
\description{
Function to extract egoitz_colors colors as hexadecimal codes
}
