% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_contingencia.R
\name{tabla_contingencia}
\alias{tabla_contingencia}
\title{Tabla de Contingencia}
\usage{
tabla_contingencia(var1, var2)
}
\arguments{
\item{var1}{Variable que se quiera introducir para formar una tabla de contingencia}

\item{var2}{Variable que se quiera introducir para formar una tabla de contingencia}
}
\description{
Tabla de Contingencia
}
