% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_nominal_analisis.R
\name{variable_nominal_analisis}
\alias{variable_nominal_analisis}
\title{Análisis de variable nominal}
\usage{
variable_nominal_analisis(variable, frecuencia = TRUE, grafico = TRUE)
}
\description{
Análisis de variable nominal
}
